Split model.hpp into eq_node.hpp and glue_node.hpp
